moves label list out of cards since its used elsewhere
