+++ src/components/activation-displays/action-renderer.tsx
